+++ src/mask/background-worker/plugin-register.ts
