+++ src/mask/typed-message/registry.ts
